==> code/02d_metrics_carbonplan.ipynb <==
# NOTE: may need to run the following command in terminal to install regionmask:
# mamba install -c conda-forge regionmask cartopy pygeos
# mamba install -c anaconda cryptography==38.0.4

import getpass
import io
import os

import azure.storage.blob
import intake
import numpy as np
import pandas as pd

# import regionmask
import xarray as xr
import zarr

### Preliminaries

######################
# Azure blob storage
######################
# connection string (from azure web login, select your storage account, then "Access keys")
connection_string = getpass.getpass()


# format storage
container_client = azure.storage.blob.ContainerClient.from_connection_string(
    connection_string, container_name="mpctransfer"
)

###################
# Models
###################
from utils import deepsdbc_dict, gardsv_ssp_dict, gardsv_var_dict

#################
# Data access
#################

# Complete catalog
cat = intake.open_esm_datastore(
    "https://cpdataeuwest.blob.core.windows.net/cp-cmip/version1/catalogs/global-downscaled-cmip6.json"
)


# function to grab all variables and SSPs for singe model/method
def grab_model(method, model, scenarios):
    # Search catalogue for method, model, all SSPs, all vars
    dsets = cat.search(method=method, source_id=model, experiment_id=scenarios, timescale="day").to_dataset_dict()

    # Concat along SSP dimension and return
    ds_ssp = []
    for key in list(dsets.keys()):
        # Get single ssp file
        ds_tmp = dsets[key]

        # Add ssp dimension
        if "experiment_id" in ds_tmp.attrs:
            ds_tmp = ds_tmp.assign_coords(ssp=ds_tmp.attrs["experiment_id"])
        else:
            ds_tmp = ds_tmp.assign_coords(ssp=ds_tmp.attrs["intake_esm_attrs:experiment_id"])

        # For some models/methods we are missing precip
        # so need to fill with NaNs
        if "pr" not in ds_tmp.data_vars:
            ds_tmp["pr"] = xr.full_like(ds_tmp["tasmax"], np.NaN)
            ds_tmp["pr"].attrs = {"units": "NaN"}

        # Append
        ds_ssp.append(ds_tmp)

    # Rechunk for faster compuations
    ds_out = xr.concat(ds_ssp, dim="ssp")
    ds_out = ds_out.chunk(
        {"ssp": 1, "lat": 360, "lon": 360, "time": tuple(ds_out.time.groupby(ds_out.time.dt.year).count().to_numpy())}
    )
    return ds_out

# Function for longest consecutive spell if needed
def n_longest_consecutive(ds, dim="time"):
    ds = ds.cumsum(dim=dim) - ds.cumsum(dim=dim).where(ds == 0).ffill(dim=dim).fillna(0)
    return ds.max(dim=dim)

###############################
# Set paths
# UPDATE THIS FOR REPRODUCTION
###############################
quantile_path = "/home/jovyan/PlanetaryComputerExamples/DownscalingComparison/lafferty-sriver_inprep_tbd/data/quantiles/"  # location of GMFD, ERA5 quantiles

##################
# Convert units
##################
def convert_units(ds):
    if "tasmax" in ds.data_vars and ds.tasmax.attrs["units"] == "K":
        ds["tasmax"] = ds["tasmax"] - 273.15
    if "tasmin" in ds.data_vars and ds.tasmin.attrs["units"] == "K":
        ds["tasmin"] = ds["tasmin"] - 273.15
    if "pr" in ds.data_vars and ds.pr.attrs["units"] == "kg m-2 s-1":
        ds["pr"] = ds["pr"] * 86400

    return ds

#########
# Dask
#########
import dask_gateway

gateway = dask_gateway.Gateway()

# cluster options
cluster_options = gateway.cluster_options()
cluster_options["worker_memory"] = 24
cluster_options["worker_cores"] = 1

# start cluster
cluster = gateway.new_cluster(cluster_options)
client = cluster.get_client()
cluster.scale(39)

# dashboard link
print(cluster.dashboard_link)

# GARD-SV

# Preliminaries
method = "GARD-SV"
models = list(gardsv_ssp_dict.keys())

assert models == list(gardsv_var_dict.keys())

## Simple metrics

### Annual avgs

# loop through models: RUNTIME IS AROUND 5 MINS PER MODEL WITH 40 DASK WORKERS
for model in models:
    # Grab model
    ds = grab_model(method, model, gardsv_ssp_dict[model])

    ds = convert_units(ds)
    ds["tas"] = (ds["tasmin"] + ds["tasmax"]) / 2.0

    # Annual averages
    ds_final = ds.resample(time="1Y").mean()

    # storage options
    ds_final = ds_final.chunk({"ssp": 1, "time": 30, "lat": 720, "lon": 1440})

    compressor = zarr.Blosc(cname="zstd", clevel=3)
    encoding = {vname: {"compressor": compressor} for vname in ds_final.data_vars}

    azure_prefix = "carbonplan/" + method + "/avg/" + model
    store = zarr.ABSStore(client=container_client, prefix=azure_prefix)

    # store
    ds_final.to_zarr(store=store, encoding=encoding, consolidated=True, mode="w")

    print(model)

### 1-day max

# loop through models: RUNTIME IS AROUND 15 MINS PER MODEL WITH 40 DASK WORKERS
for model in models:
    # Grab model
    ds = grab_model(method, model, gardsv_ssp_dict[model])

    ds = convert_units(ds)
    ds["tas"] = (ds["tasmin"] + ds["tasmax"]) / 2.0

    # Annual maxs
    ds_final = ds.resample(time="1Y").max()

    # storage options
    ds_final = ds_final.chunk({"ssp": 1, "time": 30, "lat": 720, "lon": 1440})

    compressor = zarr.Blosc(cname="zstd", clevel=3)
    encoding = {vname: {"compressor": compressor} for vname in ds_final.data_vars}

    azure_prefix = "carbonplan/" + method + "/max/" + model
    store = zarr.ABSStore(client=container_client, prefix=azure_prefix)

    # store
    ds_final.to_zarr(store=store, encoding=encoding, consolidated=True, mode="w")

    print(model)

### 5-day max

import xclim

xclim.set_options(cf_compliance="log");

# loop through models: RUNTIME IS AROUND 15 MINS PER MODEL WITH 40 DASK WORKERS
for model in models:
    # Grab model
    ds = grab_model(method, model, gardsv_ssp_dict[model])
    ds = convert_units(ds)
    ds["tas"] = (ds["tasmin"] + ds["tasmax"]) / 2.0
    ds["pr"].attrs["units"] = "mm/day"
    ds = ds.isel(member_id=0).drop("member_id")

    # Compute
    ds_RX5day = xclim.indicators.icclim.RX5day(ds=ds[["pr"]], freq="Y")
    ds_temp5day = ds[["tas", "tasmin", "tasmax"]].rolling(time=5).mean().resample(time="1Y").max()

    # Storage options
    ds_final = xr.merge([ds_temp5day, ds_RX5day])
    ds_final = ds_final.chunk({"ssp": 1, "time": 30, "lat": 721, "lon": 1440})

    compressor = zarr.Blosc(cname="zstd", clevel=3)
    encoding = {vname: {"compressor": compressor} for vname in ds_final.data_vars}

    azure_prefix = "carbonplan/" + method + "/max5d/" + model
    store = zarr.ABSStore(client=container_client, prefix=azure_prefix)

    # store
    ds_final.to_zarr(store=store, encoding=encoding, consolidated=True, mode="w")

    print(model)

### Dry days

# Function for longest consecutive spell if needed
def n_longest_consecutive(ds):
    ds_out = ds.cumsum(dim="time") - ds.cumsum(dim="time").where(ds == 0).ffill(dim="time").fillna(0)
    return ds_out.max(dim="time")

# loop through models: RUNTIME IS AROUND 15 MINS PER MODEL WITH 40 DASK WORKERS
for model in models:
    # Grab model
    ds = grab_model(method, model, gardsv_ssp_dict[model])
    ds = convert_units(ds)

    if ds["pr"].attrs["units"] == "NaN":
        print(model + " contains no precip")
        continue

    # Select only precip
    ds = ds.drop(["tasmin", "tasmax"])

    # Compute
    # Number of dry days
    ds_tmp_0 = (ds == 0.0).resample(time="1Y").sum()  # 0mm
    ds_tmp_1 = (ds < 1.0).resample(time="1Y").sum()  # less than 1mm
    # Longest sonsecutive dry day streak
    ds_tmp_0c = (ds == 0.0).resample(time="1Y").apply(n_longest_consecutive)  # 0mm longest consecutive
    ds_tmp_1c = (ds < 1.0).resample(time="1Y").apply(n_longest_consecutive)  # less than 1mm longest consecutive
    # Merge
    ds_final = xr.merge(
        [
            ds_tmp_0.rename({"pr": "count_eq_0"}),
            ds_tmp_0c.rename({"pr": "streak_eq_0"}),
            ds_tmp_1.rename({"pr": "count_lt_1"}),
            ds_tmp_1c.rename({"pr": "streak_lt_1"}),
        ]
    )

    # storage options
    ds_final = ds_final.chunk({"ssp": 1, "time": 30, "lat": 720, "lon": 1440})

    compressor = zarr.Blosc(cname="zstd", clevel=3)
    encoding = {vname: {"compressor": compressor} for vname in ds_final.data_vars}

    azure_prefix = "carbonplan/" + method + "/dry/" + model
    store = zarr.ABSStore(client=container_client, prefix=azure_prefix)

    # store
    ds_final.to_zarr(store=store, encoding=encoding, consolidated=True, mode="w")

    print(model)

## Less simple metrics

### Wet days

%%time

# Load quantiles
ds_q_era5 = xr.open_dataset(quantile_path + "era5_precip_quantiles_gardsv.nc")
ds_q_gmfd = xr.open_dataset(quantile_path + "gmfd_precip_quantiles_gardsv.nc")

# loop through models: RUNTIME IS AROUND 20 MINS PER MODEL WITH 40 DASK WORKERS
for model in models:
    # Grab model
    ds = grab_model(method, model, gardsv_ssp_dict[model])
    ds = convert_units(ds)

    if ds["pr"].attrs["units"] == "NaN":
        print(model + " contains no precip")
        continue

    # Select only precip
    ds = ds.drop(["tasmin", "tasmax"])

    ## Calculate metrics
    var_id = "pr"
    ds_tmp_out = []
    for rp in ["q99", "rp10"]:
        # Get above/below binary
        ds_tmp_q_era5 = ds[var_id] > ds_q_era5[var_id + "_" + rp]
        ds_tmp_q_gmfd = ds[var_id] > ds_q_gmfd[var_id + "_" + rp]

        # Count of hot days
        ds_tmp_q_era5_count = ds_tmp_q_era5.resample(time="1Y").sum()
        ds_tmp_out.append(xr.Dataset({var_id + "_" + rp + "era5_count": ds_tmp_q_era5_count}))
        ds_tmp_q_gmfd_count = ds_tmp_q_gmfd.resample(time="1Y").sum()
        ds_tmp_out.append(xr.Dataset({var_id + "_" + rp + "gmfd_count": ds_tmp_q_gmfd_count}))

        # Longest consecutive hot day streak
        ds_tmp_q_era5_streak = ds_tmp_q_era5.resample(time="1Y").apply(n_longest_consecutive)
        ds_tmp_out.append(xr.Dataset({var_id + "_" + rp + "era5_streak": ds_tmp_q_era5_streak}))
        ds_tmp_q_gmfd_streak = ds_tmp_q_gmfd.resample(time="1Y").apply(n_longest_consecutive)
        ds_tmp_out.append(xr.Dataset({var_id + "_" + rp + "gmfd_streak": ds_tmp_q_gmfd_streak}))

    # Merge metrics and append
    ds_final = xr.merge(ds_tmp_out)

    # storage options
    ds_final = ds_final.chunk({"ssp": 1, "time": 20, "lat": 720, "lon": 1440})

    compressor = zarr.Blosc(cname="zstd", clevel=3)
    encoding = {vname: {"compressor": compressor} for vname in ds_final.data_vars}

    azure_prefix = "carbonplan/" + method + "/wet/" + model
    store = zarr.ABSStore(client=container_client, prefix=azure_prefix)

    # store
    ds_final.to_zarr(store=store, encoding=encoding, consolidated=True, mode="w")
    print(model)

### Hot days

%%time

# Load quantiles
ds_q_era5 = xr.open_dataset(quantile_path + "era5_temperature_quantiles_gardsv.nc")
ds_q_gmfd = xr.open_dataset(quantile_path + "gmfd_temperature_quantiles_gardsv.nc")

# loop through models: RUNTIME IS AROUND 10 MINS PER MODEL WITH 40 DASK WORKERS
for model in models:
    # Grab model
    ds = grab_model(method, model, gardsv_ssp_dict[model])
    ds = convert_units(ds)
    ds["tas"] = (ds["tasmax"] + ds["tasmin"]) / 2.0

    # Drop precip
    ds = ds.drop(["pr"])

    ## Calculate metrics
    ds_tmp_final = []
    for var_id in ["tasmin", "tasmax", "tas"]:
        ds_tmp_out = []
        for rp in ["q99", "rp10"]:
            # Get above/below binary
            ds_tmp_q_era5 = ds[var_id] > ds_q_era5[var_id + "_" + rp]
            ds_tmp_q_gmfd = ds[var_id] > ds_q_gmfd[var_id + "_" + rp]

            # Count of hot days
            ds_tmp_q_era5_count = ds_tmp_q_era5.resample(time="1Y").sum()
            ds_tmp_out.append(xr.Dataset({var_id + "_" + rp + "era5_count": ds_tmp_q_era5_count}))
            ds_tmp_q_gmfd_count = ds_tmp_q_gmfd.resample(time="1Y").sum()
            ds_tmp_out.append(xr.Dataset({var_id + "_" + rp + "gmfd_count": ds_tmp_q_gmfd_count}))

            # Longest consecutive hot day streak
            ds_tmp_q_era5_streak = ds_tmp_q_era5.resample(time="1Y").apply(n_longest_consecutive)
            ds_tmp_out.append(xr.Dataset({var_id + "_" + rp + "era5_streak": ds_tmp_q_era5_streak}))
            ds_tmp_q_gmfd_streak = ds_tmp_q_gmfd.resample(time="1Y").apply(n_longest_consecutive)
            ds_tmp_out.append(xr.Dataset({var_id + "_" + rp + "gmfd_streak": ds_tmp_q_gmfd_streak}))

        # Merge RPs and append
        ds_out = xr.merge(ds_tmp_out)
        ds_tmp_final.append(ds_out)

    # Merge variables
    ds_final = xr.merge(ds_tmp_final)

    # storage options
    ds_final = ds_final.chunk({"ssp": 1, "time": 20, "lat": 720, "lon": 1440})

    compressor = zarr.Blosc(cname="zstd", clevel=3)
    encoding = {vname: {"compressor": compressor} for vname in ds_final.data_vars}

    azure_prefix = "carbonplan/" + method + "/hot/" + model
    store = zarr.ABSStore(client=container_client, prefix=azure_prefix)

    # store
    ds_final.to_zarr(store=store, encoding=encoding, consolidated=True, mode="w")
    print(model)

## Multivariate metrics

### Hot and dry days

%%time

# Load quantiles
ds_q_era5 = xr.open_dataset(quantile_path + "era5_temperature_quantiles_gardsv.nc")
ds_q_gmfd = xr.open_dataset(quantile_path + "gmfd_temperature_quantiles_gardsv.nc")

# loop through models: RUNTIME IS AROUND 10 MINS PER MODEL WITH 40 DASK WORKERS
for model in models:
    # Grab model
    ds = grab_model(method, model, gardsv_ssp_dict[model])
    ds = convert_units(ds)

    if ds["pr"].attrs["units"] == "NaN":
        print(model + " contains no precip")
        continue

    # Drop tasmin
    ds = ds.drop(["tasmin"])

    ## Calculate metrics
    ds_tmp_out = []
    for rp in ["q99", "rp10"]:
        # Get above/below binary
        ds_tmp_q_gmfd = (ds["tasmax"] > ds_q_gmfd["tasmax_" + rp]) & (ds["pr"] < 1.0)
        ds_tmp_q_era5 = (ds["tasmax"] > ds_q_era5["tasmax_" + rp]) & (ds["pr"] < 1.0)

        # Count of hot+dry days
        ds_tmp_q_era5_count = ds_tmp_q_era5.resample(time="1Y").sum()
        ds_tmp_out.append(xr.Dataset({"hotdry_" + rp + "era5_count": ds_tmp_q_era5_count}))
        ds_tmp_q_gmfd_count = ds_tmp_q_gmfd.resample(time="1Y").sum()
        ds_tmp_out.append(xr.Dataset({"hotdry_" + rp + "gmfd_count": ds_tmp_q_gmfd_count}))

        # Longest consecutive hot+dry day streak
        ds_tmp_q_era5_streak = ds_tmp_q_era5.resample(time="1Y").apply(n_longest_consecutive)
        ds_tmp_out.append(xr.Dataset({"hotdry_" + rp + "era5_streak": ds_tmp_q_era5_streak}))
        ds_tmp_q_gmfd_streak = ds_tmp_q_gmfd.resample(time="1Y").apply(n_longest_consecutive)
        ds_tmp_out.append(xr.Dataset({"hotdry_" + rp + "gmfd_streak": ds_tmp_q_gmfd_streak}))

    # Merge metrics and append
    ds_final = xr.merge(ds_tmp_out)

    # storage options
    ds_final = ds_final.chunk({"ssp": 1, "time": 10, "lat": 720, "lon": 1440})

    compressor = zarr.Blosc(cname="zstd", clevel=3)
    encoding = {vname: {"compressor": compressor} for vname in ds_final.data_vars}

    azure_prefix = "carbonplan/" + method + "/hotdry/" + model
    store = zarr.ABSStore(client=container_client, prefix=azure_prefix)

    # store
    ds_final.to_zarr(store=store, encoding=encoding, consolidated=True, mode="w")
    print(model)

# DeepSD-BC

# Preliminaries
method = "DeepSD-BC"
models = list(deepsdbc_dict.keys())

## Simple metrics

### Annual avgs

# loop through models: RUNTIME IS AROUND 20 MINS PER MODEL WITH 40 DASK WORKERS
for model in models[1:]:
    # Grab model
    ds = grab_model(method, model, [ssp for ssp in list(deepsdbc_dict[model].keys())])

    ds = convert_units(ds)
    ds["tas"] = (ds["tasmin"] + ds["tasmax"]) / 2.0

    # Annual averages
    ds_final = ds.resample(time="1Y").mean()

    # storage options
    ds_final = ds_final.chunk({"ssp": 1, "time": 30, "lat": 720, "lon": 1440})

    compressor = zarr.Blosc(cname="zstd", clevel=3)
    encoding = {vname: {"compressor": compressor} for vname in ds_final.data_vars}

    azure_prefix = "carbonplan/" + method + "/avg/" + model
    store = zarr.ABSStore(client=container_client, prefix=azure_prefix)

    # store
    ds_final.to_zarr(store=store, encoding=encoding, consolidated=True, mode="w")

    print(model)

### 1-day max

# loop through models: RUNTIME IS AROUND 15 MINS PER MODEL WITH 40 DASK WORKERS
for model in models:
    # Grab model
    ds = grab_model(method, model, [ssp for ssp in list(deepsdbc_dict[model].keys())])

    ds = convert_units(ds)
    ds["tas"] = (ds["tasmin"] + ds["tasmax"]) / 2.0

    # Annual maxs
    ds_final = ds.resample(time="1Y").max()

    # storage options
    ds_final = ds_final.chunk({"ssp": 1, "time": 30, "lat": 720, "lon": 1440})

    compressor = zarr.Blosc(cname="zstd", clevel=3)
    encoding = {vname: {"compressor": compressor} for vname in ds_final.data_vars}

    azure_prefix = "carbonplan/" + method + "/max/" + model
    store = zarr.ABSStore(client=container_client, prefix=azure_prefix)

    # store
    ds_final.to_zarr(store=store, encoding=encoding, consolidated=True, mode="w")

    print(model)

### 5-day max

import xclim

xclim.set_options(cf_compliance="log");

# loop through models: RUNTIME IS AROUND 15 MINS PER MODEL WITH 40 DASK WORKERS
for model in models:
    # Grab model
    ds = grab_model(method, model, [ssp for ssp in list(deepsdbc_dict[model].keys())])
    ds = convert_units(ds)
    ds["tas"] = (ds["tasmin"] + ds["tasmax"]) / 2.0
    ds["pr"].attrs["units"] = "mm/day"
    ds = ds.isel(member_id=0).drop("member_id")

    # Compute
    ds_RX5day = xclim.indicators.icclim.RX5day(ds=ds[["pr"]], freq="Y")
    ds_temp5day = ds[["tas", "tasmin", "tasmax"]].rolling(time=5).mean().resample(time="1Y").max()

    # Storage options
    ds_final = xr.merge([ds_temp5day, ds_RX5day])
    ds_final = ds_final.chunk({"ssp": 1, "time": 30, "lat": 721, "lon": 1440})

    compressor = zarr.Blosc(cname="zstd", clevel=3)
    encoding = {vname: {"compressor": compressor} for vname in ds_final.data_vars}

    azure_prefix = "carbonplan/" + method + "/max5d/" + model
    store = zarr.ABSStore(client=container_client, prefix=azure_prefix)

    # store
    ds_final.to_zarr(store=store, encoding=encoding, consolidated=True, mode="w")

    print(model)

### Dry days

# Function for longest consecutive spell if needed
def n_longest_consecutive(ds):
    ds_out = ds.cumsum(dim="time") - ds.cumsum(dim="time").where(ds == 0).ffill(dim="time").fillna(0)
    return ds_out.max(dim="time")

# loop through models: RUNTIME IS AROUND 15 MINS PER MODEL WITH 40 DASK WORKERS
for model in models:
    # Grab model
    ds = grab_model(
        method, model, [ssp for ssp in list(deepsdbc_dict[model].keys()) if "pr" in deepsdbc_dict[model][ssp]]
    )
    ds = convert_units(ds)

    # Select only precip
    ds = ds.drop(["tasmin", "tasmax"])

    # Compute
    # Number of dry days
    ds_tmp_0 = (ds == 0.0).resample(time="1Y").sum()  # 0mm
    ds_tmp_1 = (ds < 1.0).resample(time="1Y").sum()  # less than 1mm
    # Longest sonsecutive dry day streak
    ds_tmp_0c = (ds == 0.0).resample(time="1Y").apply(n_longest_consecutive)  # 0mm longest consecutive
    ds_tmp_1c = (ds < 1.0).resample(time="1Y").apply(n_longest_consecutive)  # less than 1mm longest consecutive
    # Merge
    ds_final = xr.merge(
        [
            ds_tmp_0.rename({"pr": "count_eq_0"}),
            ds_tmp_0c.rename({"pr": "streak_eq_0"}),
            ds_tmp_1.rename({"pr": "count_lt_1"}),
            ds_tmp_1c.rename({"pr": "streak_lt_1"}),
        ]
    )

    # storage options
    ds_final = ds_final.chunk({"ssp": 1, "time": 30, "lat": 720, "lon": 1440})

    compressor = zarr.Blosc(cname="zstd", clevel=3)
    encoding = {vname: {"compressor": compressor} for vname in ds_final.data_vars}

    azure_prefix = "carbonplan/" + method + "/dry/" + model
    store = zarr.ABSStore(client=container_client, prefix=azure_prefix)

    # store
    ds_final.to_zarr(store=store, encoding=encoding, consolidated=True, mode="w")

    print(model)

## Less simple metrics

### Wet days

%%time

# Load quantiles
ds_q_era5 = xr.open_dataset(quantile_path + "era5_precip_quantiles_nex-cil-deepsd.nc")
ds_q_era5["lon"] = np.where(ds_q_era5["lon"] > 180, ds_q_era5["lon"] - 360, ds_q_era5["lon"])
ds_q_era5 = ds_q_era5.sortby("lon")

ds_q_gmfd = xr.open_dataset(quantile_path + "gmfd_precip_quantiles_nex-cil-deepsd.nc")
ds_q_gmfd["lon"] = np.where(ds_q_gmfd["lon"] > 180, ds_q_gmfd["lon"] - 360, ds_q_gmfd["lon"])
ds_q_gmfd = ds_q_gmfd.sortby("lon")

# Loop through models: RUNTIME IS AROUND 15 MINS PER MODEL WITH 40 DASK WORKERS
for model in models:
    # Grab model
    ds = grab_model(
        method, model, [ssp for ssp in list(deepsdbc_dict[model].keys()) if "pr" in deepsdbc_dict[model][ssp]]
    )
    ds = convert_units(ds)

    if ds["pr"].attrs["units"] == "NaN":
        print(model + " contains no precip")
        continue

    # Select only precip
    ds = ds.drop(["tasmin", "tasmax"])

    ## Calculate metrics
    var_id = "pr"
    ds_tmp_out = []
    for rp in ["q99", "rp10"]:
        # Get above/below binary
        ds_tmp_q_era5 = ds[var_id] > ds_q_era5[var_id + "_" + rp]
        ds_tmp_q_gmfd = ds[var_id] > ds_q_gmfd[var_id + "_" + rp]

        # Count of hot days
        ds_tmp_q_era5_count = ds_tmp_q_era5.resample(time="1Y").sum()
        ds_tmp_out.append(xr.Dataset({var_id + "_" + rp + "era5_count": ds_tmp_q_era5_count}))
        ds_tmp_q_gmfd_count = ds_tmp_q_gmfd.resample(time="1Y").sum()
        ds_tmp_out.append(xr.Dataset({var_id + "_" + rp + "gmfd_count": ds_tmp_q_gmfd_count}))

        # Longest consecutive hot day streak
        ds_tmp_q_era5_streak = ds_tmp_q_era5.resample(time="1Y").apply(n_longest_consecutive)
        ds_tmp_out.append(xr.Dataset({var_id + "_" + rp + "era5_streak": ds_tmp_q_era5_streak}))
        ds_tmp_q_gmfd_streak = ds_tmp_q_gmfd.resample(time="1Y").apply(n_longest_consecutive)
        ds_tmp_out.append(xr.Dataset({var_id + "_" + rp + "gmfd_streak": ds_tmp_q_gmfd_streak}))

    # Merge metrics and append
    ds_final = xr.merge(ds_tmp_out)

    # storage options
    ds_final = ds_final.chunk({"ssp": 1, "time": 20, "lat": 720, "lon": 1440})

    compressor = zarr.Blosc(cname="zstd", clevel=3)
    encoding = {vname: {"compressor": compressor} for vname in ds_final.data_vars}

    azure_prefix = "carbonplan/" + method + "/wet/" + model
    store = zarr.ABSStore(client=container_client, prefix=azure_prefix)

    # store
    ds_final.to_zarr(store=store, encoding=encoding, consolidated=True, mode="w")
    print(model)

### Hot days

%%time

# Load quantiles
ds_q_era5 = xr.open_dataset(quantile_path + "era5_temperature_quantiles_nex-cil-deepsd.nc")
ds_q_era5["lon"] = np.where(ds_q_era5["lon"] > 180, ds_q_era5["lon"] - 360, ds_q_era5["lon"])
ds_q_era5 = ds_q_era5.sortby("lon")

ds_q_gmfd = xr.open_dataset(quantile_path + "gmfd_temperature_quantiles_nex-cil-deepsd.nc")
ds_q_gmfd["lon"] = np.where(ds_q_gmfd["lon"] > 180, ds_q_gmfd["lon"] - 360, ds_q_gmfd["lon"])
ds_q_gmfd = ds_q_gmfd.sortby("lon")

# Loop through models: RUNTIME IS AROUND 10 MINS PER MODEL WITH 40 DASK WORKERS
for model in models:
    # Grab model
    ds = grab_model(method, model, list(deepsdbc_dict[model].keys()))
    ds = convert_units(ds)
    ds["tas"] = (ds["tasmax"] + ds["tasmin"]) / 2.0

    # Drop precip
    ds = ds.drop(["pr"])

    ## Calculate metrics
    ds_tmp_final = []
    for var_id in ["tasmin", "tasmax", "tas"]:
        ds_tmp_out = []
        for rp in ["q99", "rp10"]:
            # Get above/below binary
            ds_tmp_q_era5 = ds[var_id] > ds_q_era5[var_id + "_" + rp]
            ds_tmp_q_gmfd = ds[var_id] > ds_q_gmfd[var_id + "_" + rp]

            # Count of hot days
            ds_tmp_q_era5_count = ds_tmp_q_era5.resample(time="1Y").sum()
            ds_tmp_out.append(xr.Dataset({var_id + "_" + rp + "era5_count": ds_tmp_q_era5_count}))
            ds_tmp_q_gmfd_count = ds_tmp_q_gmfd.resample(time="1Y").sum()
            ds_tmp_out.append(xr.Dataset({var_id + "_" + rp + "gmfd_count": ds_tmp_q_gmfd_count}))

            # Longest consecutive hot day streak
            ds_tmp_q_era5_streak = ds_tmp_q_era5.resample(time="1Y").apply(n_longest_consecutive)
            ds_tmp_out.append(xr.Dataset({var_id + "_" + rp + "era5_streak": ds_tmp_q_era5_streak}))
            ds_tmp_q_gmfd_streak = ds_tmp_q_gmfd.resample(time="1Y").apply(n_longest_consecutive)
            ds_tmp_out.append(xr.Dataset({var_id + "_" + rp + "gmfd_streak": ds_tmp_q_gmfd_streak}))

        # Merge RPs and append
        ds_out = xr.merge(ds_tmp_out)
        ds_tmp_final.append(ds_out)

    # Merge variables
    ds_final = xr.merge(ds_tmp_final)

    # storage options
    ds_final = ds_final.chunk({"ssp": 1, "time": 20, "lat": 720, "lon": 1440})

    compressor = zarr.Blosc(cname="zstd", clevel=3)
    encoding = {vname: {"compressor": compressor} for vname in ds_final.data_vars}

    azure_prefix = "carbonplan/" + method + "/hot/" + model
    store = zarr.ABSStore(client=container_client, prefix=azure_prefix)

    # store
    ds_final.to_zarr(store=store, encoding=encoding, consolidated=True, mode="w")
    print(model)

## Multivariate metrics

### Hot and dry days

%%time

# Load quantiles
ds_q_era5 = xr.open_dataset(quantile_path + "era5_temperature_quantiles_nex-cil-deepsd.nc")
ds_q_era5["lon"] = np.where(ds_q_era5["lon"] > 180, ds_q_era5["lon"] - 360, ds_q_era5["lon"])
ds_q_era5 = ds_q_era5.sortby("lon")

ds_q_gmfd = xr.open_dataset(quantile_path + "gmfd_temperature_quantiles_nex-cil-deepsd.nc")
ds_q_gmfd["lon"] = np.where(ds_q_gmfd["lon"] > 180, ds_q_gmfd["lon"] - 360, ds_q_gmfd["lon"])
ds_q_gmfd = ds_q_gmfd.sortby("lon")

# loop through models: RUNTIME IS AROUND 10 MINS PER MODEL WITH 40 DASK WORKERS
for model in models:
    # Grab model
    ds = grab_model(
        method, model, [ssp for ssp in list(deepsdbc_dict[model].keys()) if "pr" in deepsdbc_dict[model][ssp]]
    )
    ds = convert_units(ds)

    # Drop tasmin
    ds = ds.drop(["tasmin"])

    ## Calculate metrics
    ds_tmp_out = []
    for rp in ["q99", "rp10"]:
        # Get above/below binary
        ds_tmp_q_gmfd = (ds["tasmax"] > ds_q_gmfd["tasmax_" + rp]) & (ds["pr"] < 1.0)
        ds_tmp_q_era5 = (ds["tasmax"] > ds_q_era5["tasmax_" + rp]) & (ds["pr"] < 1.0)

        # Count of hot+dry days
        ds_tmp_q_era5_count = ds_tmp_q_era5.resample(time="1Y").sum()
        ds_tmp_out.append(xr.Dataset({"hotdry_" + rp + "era5_count": ds_tmp_q_era5_count}))
        ds_tmp_q_gmfd_count = ds_tmp_q_gmfd.resample(time="1Y").sum()
        ds_tmp_out.append(xr.Dataset({"hotdry_" + rp + "gmfd_count": ds_tmp_q_gmfd_count}))

        # Longest consecutive hot+dry day streak
        ds_tmp_q_era5_streak = ds_tmp_q_era5.resample(time="1Y").apply(n_longest_consecutive)
        ds_tmp_out.append(xr.Dataset({"hotdry_" + rp + "era5_streak": ds_tmp_q_era5_streak}))
        ds_tmp_q_gmfd_streak = ds_tmp_q_gmfd.resample(time="1Y").apply(n_longest_consecutive)
        ds_tmp_out.append(xr.Dataset({"hotdry_" + rp + "gmfd_streak": ds_tmp_q_gmfd_streak}))

    # Merge metrics and append
    ds_final = xr.merge(ds_tmp_out)

    # storage options
    ds_final = ds_final.chunk({"ssp": 1, "time": 20, "lat": 720, "lon": 1440})

    compressor = zarr.Blosc(cname="zstd", clevel=3)
    encoding = {vname: {"compressor": compressor} for vname in ds_final.data_vars}

    azure_prefix = "carbonplan/" + method + "/hotdry/" + model
    store = zarr.ABSStore(client=container_client, prefix=azure_prefix)

    # store
    ds_final.to_zarr(store=store, encoding=encoding, consolidated=True, mode="w")
    print(model)
